refactor: move balances-activity to feature
