+++ src/pages/home/home.tsx
@@ -3,7 +3,7 @@ import { Outlet } from 'react-router-dom';
 import { Box, Stack } from '@stacks/ui';
 import { PopupContainer } from '@components/popup/container';
 import { Header } from '@components/header';
-import { BalancesAndActivity } from '@components/popup/balances-and-activity';
+import { BalancesAndActivity } from '@features/balances-and-activity';
 import { UserAccount } from '@pages/home/components/user-area';
 import { HomeActions } from '@pages/home/components/actions';
 import { HiroMessages } from '@features/hiro-messages/hiro-messages';
